+++ frontend/src/component/user/UserProfile/index.tsx
@@ -1,14 +1,21 @@
-import UserProfile from './UserProfile.tsx';
 import { useAuthUser } from 'hooks/api/getters/useAuth/useAuthUser';
+import { useUiFlag } from 'hooks/useUiFlag';
+import { UserProfile } from './UserProfile.tsx';
+import { LegacyUserProfile } from './LegacyUserProfile.tsx';
 
 const UserProfileContainer = () => {
     const { user } = useAuthUser();
+    const newProfileDropdown = useUiFlag('newProfileDropdown');
 
     if (!user) {
         return null;
     }
 
-    return <UserProfile profile={user} />;
+    return newProfileDropdown ? (
+        <UserProfile profile={user} />
+    ) : (
+        <LegacyUserProfile profile={user} />
+    );
 };
 
 export default UserProfileContainer;
